conan/requirements.txt
pyldap>=2.4.28
configparser>=3.5.0